Add files via upload
Uploaded the codes made by Pipple. The code 'PDF_Text_concerter' is an older script, our focus the last weeks was on 'DocumentAnalyzer' where the converted .pdf files to .txt files are analyzed on data, disaster, location and victims. The script reads the .txt files from the 'Input' folder and creates Excel files which are stored in the 'Output' folder.